fix some bugs

+++ beverage_shop_database.sql
@@ -34,7 +34,7 @@ CREATE TABLE user_info
     user_id INT PRIMARY KEY AUTO_INCREMENT,
     real_name VARCHAR(255) NULL, -- 真实姓名
     gender INT NULL DEFAULT NULL, -- 0男，1女, NULL未知，默认值为NULL
-    telephone INT NULL,
+    telephone BIGINT NULL,
     delete_flag INT DEFAULT  0, -- 0正常，1已被删除
 
     -- staff_specific_field
